Added Allocated Cluster for running jobs & Runned On Cluster for the completed jobs

+++ frontend/src/pages/Management/Clusters/DeleteClusterForm.js
@@ -7,38 +7,45 @@ const DeleteClusterForm = ({ cluster, closeForm, deleteCluster }) => {
     // Function to handle the delete operation
     const handleDelete = async () => {
         try {
-            if (cluster.poolConnectedTo != null) {
-                const response = await fetch(`http://localhost:3000/management/pools/removeClusterFromPoolById/${cluster.poolConnectedTo}`, {
-                    method: 'PUT',
+            if (cluster.clusterStatus === 'Available') {
+                if (cluster.poolConnectedTo != null) {
+                    const response = await fetch(`http://localhost:3000/management/pools/removeClusterFromPoolById/${cluster.poolConnectedTo}`, {
+                        method: 'PUT',
+                        headers: {
+                            'Content-Type': 'application/json'
+                        },
+                        body: JSON.stringify({ cluster })
+                    });
+
+                    if (!response.ok) {
+                        throw new Error(`HTTP error! status: ${response.status}`);
+                    }
+                }
+
+                // Sending a DELETE request to the cluster
+                const response = await fetch(`http://localhost:3000/management/clusters/deleteClusterById/${cluster.clusterId}`, {
+                    method: 'DELETE',
                     headers: {
                         'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify({ cluster })
+                    }
                 });
 
                 if (!response.ok) {
                     throw new Error(`HTTP error! status: ${response.status}`);
                 }
+
+                const result = await response.json();
+                console.log(result.message); // Log the success message (optional)
+
+                // deleteCluster(cluster.clusterId); // Update the parent component's state by removing the cluster
+                deleteCluster(cluster); // Update the parent component's state by removing the cluster
+                closeForm(); // Close the form after deletion
             }
-
-            // Sending a DELETE request to the cluster
-            const response = await fetch(`http://localhost:3000/management/clusters/deleteClusterById/${cluster.clusterId}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
+            else {
+                alert('Please try again later. The cluster is running or unavailable.');
+                return; // Exit the function without submitting the form
             }
-
-            const result = await response.json();
-            console.log(result.message); // Log the success message (optional)
-
-            // deleteCluster(cluster.clusterId); // Update the parent component's state by removing the cluster
-            deleteCluster(cluster); // Update the parent component's state by removing the cluster
-            closeForm(); // Close the form after deletion
+                
         } catch (error) {
             console.error('Error deleting cluster:', error);
         }
